Move tracker notification settings management outside of the Tracker class
This contribution does not change everything at the functionnal level,
everything is working (or is broken) as before. This is only a technical
change to be able to rework the notification management in saner conditions.

Reviewers, do not forget to redeploy your nginx configuration to test this
contribution.

This is part of story #11104: monitor a tracker by email

Change-Id: Ia749aa9ef7bf1b38297a8db5d65192d3ce82c43e