Convert BuildItemMappingVisitorTest and some others
It exposes the last version of simpletest2mockery that support a nice conversion
between the 2 mocking system.

Alas there is one pitfall with mockery: default values.
With simpletest everything that was not explicitely defined got false
so

    $dar->setReturnValue('getRow', array(30));
    $dar->setReturnValue('getRow', array(40));

Was good because we where able to to loop like while ($row = $dar->getRow())
and the third call to $dar->getRow() whould have got false and that terminate
the loop.

Mockery has a different strategy, the first value is the default value
so, with

    $dar->shouldReceive('getRow')->once()->andReturns(30);
    $dar->shouldReceive('getRow')->once()->andReturns(40);

The third call to $dar->getRow() will get 30 instead of false.

Code that looks like that will not terminate (infinite loop), you will need to
manually edit tests to have:

    $dar->shouldReceive('getRow')->once()->andReturns(30);
    $dar->shouldReceive('getRow')->once()->andReturns(40);
    $dar->shouldReceive('getRow')->once()->andReturns(false);

This is why we can't have nice things.

Part of story #11368 Convert docman tests to mockery

Change-Id: Ieee6847c813e3f6bd0ca7ff1db39324f4d7b2b09