Separate the display of the tracker notifications settings from the update
This contribution does not change everything at the functionnal level,
everything is working (or is broken) as before. This is only a technical
change to be able to rework the notification management in saner conditions.

All updates in the tracker notification settings are now protected against CSRF.

This is part of story #11104: monitor a tracker by email

Change-Id: I0ab21e44ffa304278dde935e2eb016cdf43c879b