Display snapshoted definition of execution
Part of story #11389: snapshot test executions

Go to a test definition and change many times the description. For example
"Version A", "Version B", "Version C".

Take the changeset_id of the "Version B" (Ex: 237). This changeset id is
displayed as tooltip on the artifact when you hover the followup link
icon: "Link to this followup - #237".

Create a campaign and add the test. When you try to execute the test, you
should see the latest version of the definition (Version C).

Insert in the table plugin_testmanagement_execution an entry for the given
execution id (ex: 4) and a previous changeset (ex: 237). Refresh the campaign
and go to the test. You should see a previous version of the definition
(Version B).

Update the test definition (Version D). You still see "Version B" in the TTM.

Change-Id: I2e7f6bdd337e78b6424115e931f5503fab3a1193