Only the most recent query is used to update the password check result
When a user type a password queries are sent to Tuleap to check
if the password respects the password strategy and inform the user.
Multiple queries might be sent but there is no guarantee that the
latest query will also be the latest to respond. Some queries might
be slower than other. Since the information displayed to the user
is updated whenever a result is retrieved it might happen than it
is not the more recent, the user will then get an outdated information
and be confused.

This is part of story #11182: prevent users to use a breached password

Change-Id: Ia1f97ae4d88793bbd75b42bffef04d7147b43a30