Users can be added to a DB table to ignore all notifications of a tracker
By adding an entry to the table tracker_global_notification_unsubscribers
a user can be switched to the mode "No notification at all" for a specific
tracker. Once switched to this mode:
 * the user will not receive notification from this tracker anymore
 * the user is not visible in the global email notification admin table anymore
like it's the case for user not being active
 * the user does not have the Disable/Enable notifications button per artifact

Similarly to what is done for not active users, a user in the "No notification
at all" mode can still be added in the global email notification list even if
he will not be visible nor he will receive any notifications. This behavior
will be improved in a improved contribution.

This is part of story #11104: specify how to monitor trackers

Change-Id: I626cbc5fdd404e29748357bc03392bfee02475c7