Use current version when creating a campaign
Part of story #11389: snapshot test executions

In TTM, create a new campaign with tests (all tests, or from a report).

From test def tracker, edit the definition a test to create a new version.

Go back to the campaign, the test execution is still linked to the previous
version of the test.

Change-Id: Ic2a94882de321bc63ea2cd1c56c6f8c73e04eb6e