Convert a bunch to tracker tests to simpletest 1.1.x
There is no really a common approach/group for those tests, it's just the ones
that are already compatible.

I initially intended to provide one commit for all but it's taking too much
time and rebase are hard to deal with so the sooner to master the better.

There are still 24 test suite not compatible yet (-38)

Part of story #11484 have tracker tests running with simpletest 1.1.x

Change-Id: I69dc3ed1b155562f3005888ca0eb26fcd23d25af