request #11412: Fatal error might happen when copying an artifact and a changeset is not found
In some undetermined cases exporting the XML of an artifact ends up with a fatal
error because a changeset value can not be retrieved.
The XML exporter should take into the account the Tuleap internal API (aka null
value might encountered) and do the export in a best effort mode to not raise
a fatal error.

Change-Id: Id51be7269dba8494620b4009e15edcbc09e39454