Users can change of notification mode for a tracker
When a user adds himself to the global notifications, permissions
are checked. A tracker administrator can however to remove the
the permissions check and it that case if the user switches between
the two global notification modes this settings is kept. However,
if the user disables the global notification mode then come back to it
again, the setting is lost.

This is part of story #11104: specify how to monitor trackers

Change-Id: Ie75d8ea02ef2b6e6f372ec99126d1b0de30fb59e