Move the retrieval of the remote data of a Jenkins job in a dedicated class
This contribution is purely technical, it is only an (ugly) refactoring.
The goal is to move outside the HudsonJob class the retrieval of the XML
content so that in the future it is possible to instanciate the jobs even
if takes too long. It could also be used to retrieve the XML content of
multiple jobs in parallel.

Since the existing tests for the HudsonJob class had a lot of changes,
the oppurtunity has been took to move these tests to PHPUnit.

This is part of request #11415: Displaying all CI jobs can be slow

Change-Id: I12d7c32167ff6d1f31f584a0ecd08ae382c819e3