request #11434: Dashboard of a project might not be found if the project uses a short name with upper case characters
The collation used in the DB might take the case into account but most of
the links are generated using the project short name in lower case. This contribution
ensure that when the project short name is used directly to retrieve a project
from an autocompleter or from the dashboard page, the project can be found.

Change-Id: I51d9489cb9d480b36c670ad93485ad4c9d27f889