Cypress tests should be run nightly
this is part of request #11328 move some TTM tests to cypress

E2E tests fails to often to be run for every signle patch.
A nightly run is suffisent for now.

How to test:
For now the new job TuleapNightlyTests is the same than the
jenkinsfile of this commit.
Trigger the job => it should finish with success

Correction of job will be done, once this patch will be merged

Change-Id: I1ce8f8e6a4861b2e668812724b474a717fa431a9