request #11431 gitolite admin key not in /home/codendiadm/.ssh
The location of this key depends on the system. Whether we are in el6 or el7
the home is different. The solution is harcoded in order to avoid brittle
dependencies: HOME env variable might not be defined (using sudo) and adding
a variable in a config file for something that is not variable doesn't feel
right.

Change-Id: Id7a79cc6b4d861224b59a9011e91fca468f1a137