Introduce karma-configurator
Part of request #11438 - DRY Karma installs and configurations

The first introduction of the karma configurator has been made for
pullrequest because it has a few unit test and is easier to migrate.

This is a first step, the other karma configs will be dried in the next
commits.

How to test:
- make post-checkout
- go to pullrequest/www/scripts/
- npm run test
--> Tests are launched and are all successful

- npm run coverage
--> Tests are launched, are all successful and the results are displayed
    in a browser.

- npm run watch
--> Tests are launched everytime you make a modification in a watched js
    file

Change-Id: I55bd61aadd2b392fe61adc5a114cef1b948ab51b