Git smart http routed by FrontRouter
Take the opportunity to go a little bit deeper in the FrontRouting with more
complex routes (complex regexp to indentify Smart HTTP requests) in addition
to low level permission handling.

For permission handling, we introduce a new Interface that completly disable
the existing mechanism "URLVerification". It's the common protection that
forbid access to private projects, restricted users & co.

It was previously managed by a dedicated hook in URLVerification that
"whitelisted" URLs so permissions were handled later in the stack.
With this new approach we make clear the requests that implement the new
interface are "on their own" and developer has to deal with all permission
and authentication checking manually.

Part of request #11450 FrontRouter for git plugin

Change-Id: Icb2473a5fe5241e16ec828a68c44dd50ff383d80