request #11448: Sysadmins should be able to prevent site administrators to uninstall/disable plugins from the web UI
To test you can deploy a new parameter in your local.inc file.
For example, with the following parameter you won't be able to
disable/uninstall the label or dynamic_credentials plugins:

    $sys_plugins_that_can_not_be_disabled_from_the_web_ui = 'label, dynamic_credentials';

This contribution also fixes the fact that it was possible for someone
crafting the request to disable the plugin PluginsAdministration.

Change-Id: Ib1357ad0b2d8219553e7317fdbc0d95842d7028e