Cap the maximum time given to retrieve data for a batch of jobs
In the worst scenario, the retrieval of a batch of jobs can now take
MAX_BATCH_TIME + TIMEOUT_HTTP_CLIENT - 1s.

The maximum time given to process a batch of jobs has been arbitrarily set
to 5 seconds. In the future this should be dropped in favor of multiple
HTTP queries processed in parallel.

This is part of request #11415: Displaying all CI jobs can be slow

Change-Id: I4f1fe8875fdb7ddfc6abba6aae4bb49f4c20432d