Computed value is transformed as manual value in XML archive
This is part of story #11342 have a base REST route to move id, title, description

While archiving an artifact after deletion, each computed fields that
are in computed mode in the last changeset are transformed into manual
values. Computed fields that already are in manual value are not added
in last changeset.

This allow us to store this value in archive after deleting an artifact.

Nothing changes in XML export for artifact copy and project export.

In addition, this patch provides few bug fixes for tracker export only
using the project export tool.

Change-Id: I1e6712f5a05a77d84b934aba2f3c6dea35bbc7bf