Start scrum SQL queries should be as much as possible in transactions
Transaction in Semantic was not needed at creation (because nothing to remove)
and broke the transaction that was started earlier for XML import of tracker.

Plus cardwall issued a large number of SQL writes completly out of any
transactions.

Those modifications allow to reduce the database time by 30% and the wall time
by 22% without changing the number of queries done.

Part of request #10030 start scrum timeout with slow database writes

Change-Id: I04125015be0431fc8bc38bec83ac739799f3dd0d