request #11466: Wrong level of alert in artifacts deletion configuration
> Be careful: The limit you define here is the number of artifacts users can
> delete from the REST API on a 24 hours sliding window. An artifact deletion
> is an action hardly reversible. Choose wisely.

Usage of tlp-alert-danger is inappropriate since there is no error and the
content of the message is not critical enough to justify this level. It should
be displayed as a warning instead.

Furthermore the message mix two pieces of information: one warning (deletion of
the artifact) and one info (purpose of the limit). They should be separated.

In addition icons in alerts are missing the class tlp-alert-icon. (See tlp doc
for reference)

Change-Id: I2a0afffffb6d51e5073e8946f9fe798768e37c48