Git repository browsing routed by FrontRouter
As a nice consequence it reduce the scope of the git GOD object controller.
Take also the opportunity to force usage of original, non friendly, git urls
with a redirect. The complete removal of the old URL scheme will be done in
a dedicated commit to avoid a fat commit.

Also fix pre-commit hooks so they deal with renamed files. Renamed used to
crash pre-commit hooks because more than one commit was returned.

Part of request #11450 FrontRouter for git plugin

Change-Id: Iaab61647a1ebeb44c617d5fd1f9d8ec22a1748cf