SVN users can be authenticated without a password stored in the sha512crypt format
The Tuleap.pm module is now capable to use bcrypt hashed password. This makes possible
to do SVN operations on Tuleap instances installed without the support of UNIX users
and groups.

To test this contribution, do not forget to:
 - redeploy Tuleap.pm to /usr/share/perl5/vendor_perl/Apache/Tuleap.pm
and restart Apache
 - install the new dependency

This is part of story #11435: install SVN plugin
It also solves request #11470: Not able to authenticate when doing SVN operations
with the tuleap-aio Docker image

Change-Id: I7abbc957c123675de789307c7c715ff27adea955