Move artifact with basic information
This is part of story #11342 have a base REST route to move id, title, description

The move only works with basic information:

* id
* subon
* subby

The base artifact is deleted (and archived if the plugin is enable).
The deletion limit is also used.

Next commits will introduce a global transaction, the global rank
and the title and description semantics.

Change-Id: I9a01dfc608c0525734fa130a0a54fc30e3937298