StepDefinition can only be created in project with TTM
Part of story #11390: have base of steps across the stack

*Important:* This depends on https://gerrit.tuleap.net/11382

Go to a project where TTM service is not activated. In field usage, the create
button for StepDefinition field is not present.

Bonus: Since there is no reason to have more than one StepDefinition in a
tracker, it is now unique (button will be disabled if it is already used)

Change-Id: I49ff283a5446157c4c06839fa5c1f7b432069e09