Manage cross references in step description
Part of story #11390: have base of steps across the stack

In a step description, insert a reference (e.g. art #1). Submit.

In the target reference, there is a backlink to the test definition where you
put the reference. (If this is an artifact, the backlink is displayed via the
field "Cross References")

When you display the step description, you can click on the reference.

Change-Id: I02d37a07f7b690b1341386e662e0c7872026207f