Tracker tests load bootstrap without magic constant
There is nothing that defines "TRACKER_BASE_DIR" hence when running a tracker
test suite independently on the CLI, the test crash.

This will ease the ability to run simpletest11x tests one after another.

Also add the list of test suite that are not ready yet. There are 62 test suite
not ready on 271 atm.

Part of story #11484 have tracker tests running with simpletest 1.1.x

Change-Id: I06a33a5a2954613146d34b17f3791b1c3476068d