Style steps definition in artifact view
Part of story #11390: have base of steps across the stack

Add a step in an test definition.
Directly in db, add more steps. (it is not possible to add more than one step
in UI yet)
The steps should be nicely displayed.

When you create a test without any steps, it should not create an empty step.

Change-Id: Ibdf415d85bded62677e5e67dc7f50f4c663d7f38