Display steps in TTM
Part of story #11390 have base of steps across the stack

Go to TTM » campaign » test with steps

You should see steps displayed.

Steps are considered as "notrun" for now. You can manually test the design with
the browser dev tools: on a step, change the following classes:

  {.steps-step.notrun, .steps-step-number > .tlp-badge-secondary}

to either:

- {.steps-step.passed, .steps-step-number > .tlp-badge-success}
- {.steps-step.failed, .steps-step-number > .tlp-badge-failed}
- {.steps-step.blocked, .steps-step-number > .tlp-badge-blocked}

No changes if the test does not have any steps.

Change-Id: Ia8871d69548a9608e6f22132f1befa12cd1f6132