StepExecution can only be created in project with TTM
Part of story #11390: have base of steps across the stack

Go to a project where TTM service is not activated. In field usage, the create
button for StepExecution field is not present.

Bonus: Since there is no reason to have more than one StepExecution in a
tracker, it is now unique (button will be disabled if it is already used)

Change-Id: Id89420618b7736c533d91d0691867ec68bfe4ba9