Download artifacts as single page document
Part of request #11504: Download list of artifacts as a single page document

Install the plugin and enable it. Go to a tracker report table with matching
artifacts: you can [export v] a single page document. An html file is
downloaded with each artifacts in a tlp pane: pane title is the semantic title,
pane content is the semantic description of the artifact.

Limitations (will be handled in next commits):
- there must be matching artifacts
- there should be a semantic title/description

Change-Id: If688c30561f2ea166715fd373b6b8229489727da