Prevent user to download document when no semantic
Part of request #11504: Download list of artifacts as a single page document

Given a tracker does not have title, description, or both semantics defined,
Then the option to download a single page document disappear from the menu,
And if the user forge the request, then an error message is displayed.

Note: we don't display the menu item because most people are not admin and
cannot fix the situation. This was not the case in the previous patchset because
the users can update their search criteria.

Change-Id: Ic8bea0fa8c64eb894f8f641cb59abccca7ee1cb8