request #11453: Artifact modal can be made unusable if the tracker structure is modified after an initial load
This contribution disables the agressive caching of the tracker structure in the artifact
modal to use the good enough defaults.
Currently the aggresive default can lead to issues if have used the artifact modal to edit an artifact,
modified the tracker structure (e.g. removing a field, added a value in a select box…) and then tried
to edit wgain ith the artifact modal an artifact of this tracker.

Change-Id: I7e6832425b65179b346ec18530e5ed781c2aea05