Move Tuleap admin CLI to symfony/console
This contribution is mainly a technical refactoring to
make easier to add new command to the CLI in the future.

The behavior of the existing arguments has been kept to
not break existing scripts relying on it.

This is part of request #11508: Tuleap CLI tools can reset a user password

Change-Id: Id29e26dc28c3e7827a19f5564b8c04cd32429611