Use artifact submitter instead of user that moves the artifact
This covers request #11536 Wrong user can be set a submitter while moving an artifact

The first implementation of move artifact contains a little bug.
As done for the artifact copy, the submitter was the user that ask to
move the artifact instead of the one that create it before.

REST tests are updated to handle this case.

Change-Id: I971fabcb075dc40557f4d797829a698c58540c5e