Use data attributes for html rich text editor instance
Part of story #11492: have several steps in my tests

Instead of creating hidden input text, we should be able to specify the format
used with data attributes instead of hidden inputs. Both system are supported
so there is no breaking changes.

This can be tested with https://gerrit.tuleap.net/11497

Change-Id: I56963877be74be9b26866f246e16ba4fcb9680d0