Introduce card field choice in card semantic
this is part of story #11542 Update tracker "color picker" to have beautiful colors to associate with select box values

How to test:
Go in Card semantic view.

On a tracker with no field, an empty state is displayed for car field background color section.
On a tracker with fied, a drop down should display all tracker fields
(filter on only possible field will be done in a dedicated commit).

Update/Save field for background color will be done in a dedicated commit

Change-Id: I05239bb22c9ab24e57f40248777e9837bd1085c5