Udpate steps status through REST
Part of story #11492: have several steps in my tests

How to test:
First of all make sure that the StepExecution field in test exec tracker has
update permissions for the current user.

Create a test def in a campaign. Go to the test execution. Note the id of the
execution.
In test def artifact, add a step. Submit the artifact.
In plugin_testmanagement_changeset_value_stepdef, note the id of the new step.

Go to TTM: the test def does not appear. We will fix that: edit the test title
with the TTM modal. The step should appear, in notrun state.

Go to the API explorer and issue the following:

PATCH /testmanagement_executions/<exec-id>
{
  "steps_results": [
    {"step_id": <step-id>, "status": "failed"}
  ]
}

Go back to TTM, the step is marked as failed instead of notrun.

Go to test exec artifact /plugins/tracker/?aid=<exec-id> and try to edit the
step results. A warning message tells you that this is not possible.

Change-Id: If291495dc8fc2c8d08a3e62b6bcdb061803a8117