Pull Requests references namespace is read-only for users
In every repositories when the pull request plugin is active
the refs/tlpr/* namespace can not be written by users.

The goal is to try what can go bad with PR and users manipulations.

To test you can push some refs under the namespace refs/tlpr
and then, with this contribution, force the regeneration of
the gitolite configuration.

This is part of story #11477: have a dedicated references for pullrequests

Change-Id: Idcd5c77150eb282b64eea46147190bbc32222485