Can delete steps
Part of story #11492: have several steps in my tests

*Important:* This depends on https://gerrit.tuleap.net/11537

In a test definition **with more than one step** when I edit the steps I can
delete one or more steps and submit the artifact.

*Note:* for now, we need to keep at least one step, else the field is not
changed. This will be managed in a future commit. Here we just introduce the
deletion mechanism & gettext, we don't want to clutter much the review.

Gloss and strass will come later as well.

Change-Id: Ifc1e2fe15e064ffbc72be2ba7b8115c011bd7a0d