Creation of a pull request can create a Git reference in the destination repository
If the hidden option enable_creation_git_ref_at_pr_creation is enabled, a Git reference
under the namespace refs/tlpr/ of the destination repository will be created for
each new PR.
This hidden option will be removed as soon as Tuleap is also capable to update
the reference.

The creation of the Git reference is done in such way that even if the expected
available reference is already taken, Tuleap will be able to recover and find the
next available reference in the namespace.

This is part of story #11477: have a dedicated references for pullrequests

Change-Id: I51c07b7c32704a52bf0c1080991529c8ac6565d3