Card preview display a minimal card
this is part of story #11543 Use new palette colors as background color in A.D

This patch introduce card preview in Card Smeantic
For now, the card preview display a card, without details and without background.

You should have maximum 10 cards displayed,
and card_content only display a fake id and an example of
card description.

Change-Id: Id5bc230ae75208171160d38b863c7e62cbcbb4a8