card preview: add details of card
this is part of story #11543 Use new palette colors as background color in A.D

Card preview display now the details section.
Card preview display values of an artifact of tracker.
If tracker does not have any artifacts, fields will be displayed with XXX as value

Background color will be displayed in next commit

Change-Id: Iae12c6550325e06ea127fcc52b0b0d3b57e55706