Incorrect key is returned by the PDO compat layer for DataAccessResult
When the current key is asked, the current value is returned instead of
the current index in the result. This leads to issues because it does not
match the behavior of the DataAccessResult based on the mysql_* API.

The display of the members of a ugroup in the project admin web UI is
affected by this.

This is part of story #11146: Provide PDO by default for all Tuleap DB queries

Change-Id: Iaa84e70ae78db5e44714e73705bba011d994434a