request #11594 Reduce un-necessary DB writes on start scrum
The top used queries are for prepareRanking and createNullValueOnChangeset.
As the ranking is set by XML value, there is no need to "prepare" for it.
As there are no values in those trackers, there are no need for checking which
changesets to update.

On refrence server the start scrum went from 45s down to 5.5 seconds and 350 sql
queries were saved.

Change-Id: Ib150aa190c001d46a9c136081c5ebc1ff1bc8f34