Reserve a Git reference for all existing pull requests
The Git reference is reserved and then created in the repo if something
is pushed in the branch of the pull request. If for some reasons the
creation of the Git reference fails, we will mark the Git reference as
broken and do not try to create it again as it is very unlikely to get
a working state again.

A future contribution will try to create all the missing Git references
in the repositories.

This is part of story #11477: have a dedicated references for pullrequests

Change-Id: I4501425ca8159034671042d27354dab6dd467efb