Rename Tracker::stop_notification
Part of story #11254: reduce notification sent by tracker to status change

In order to prepare introduction of a third option (notif on status change),
stop_notification is not anymore relevant.
This refactoring change in database the name of the column, and in the code.
XML import accept both names to not reject existing export. Should be
completely removed in few release.

No functional changes in the interface.

Change-Id: I39f2616d2ea3c8fe1e0a1ec9caaed155a8b84f68
TODO: use trang to generate the rng