Remove unused fgcolor from Cardwall_column
Part of story #11542 - Update tracker "color picker" to have beautiful colors to associate with select box values

The goal of this commit is to remove the unused fgcolor in
Cardwall_Columns. In the past, headers colors were displayed
on the whole header, and it was necessary to display the column
label with a light font if the background-color was dark, and
a dark font if the background-color was light for the label to be
readable. Today we only display a border at the top of the column's
header, so it needs to be removed.

How to test:
- No functional changes expected in Cardwalls in tracker reports,
  dashboards, and pv2

Change-Id: I2f9642522e3908f28a2e0e107e6df7199e061cd3