Indicate that the export is partial
Part of request #11504: Download list of artifacts as a single page document

In the download link, don't fool users by making them think that they will
download the full table as a document.

Saying "Partial export" is good enough for the customer to indicate that the
feature "is not finished". A better way to achieve this result would have be to
move this functionality in a dedicated renderer (less confusion with table
columns, plenty of space to put warnings, …), but time does not permit yet :(

Change-Id: If68683b875c2569d6a096aab0694181e0b07207d