Git operations done to display a PR are now done in the destination repository
Since the destination has now a reference to the PR, we can do all the
operations in the destination repository instead of the source repository.
This allow us to be sure that the reference will still be available in the
future.

All Git operations about merges will also be done entirely in the destination
repository in a dedicated contribution.

This is part of story #11477: have a dedicated references for pullrequests

Change-Id: I44b567fa48dd497be6f13fb4806d1e3adc505003