Add transparent color in bottom row
Part of story #11542 - Update tracker "color picker" to have beautiful colors to associate with select box values

/!\ I did the best I could to harmonize the colorpickers in the cardwall
admin. Indeed, in the tracker admin, a lot of stuff is comming from
tracker/style.scss, but this file is not included in the cardwall admin.
Plus colorpickers have a difference of width because the input holding the
value of the selected color is shown, hence the more important width.
However, I have centered the 'no color' component and I have corrected a
few little things.

How to test:
- Open the colorpicker
--> You should see a struck circled with a "No color" label at the
    bottom of the new palette

- Click on this circle
--> The color is set to transparent

- Reopen the colorpicker
--> A tick is displayed inside the circle because it is the current
   select value

Change-Id: I036c3d9c94926d62044a9aaa4cd60f491ef3ce91