Associate build status to a commit instead of a pull request
Due to the changes introduced by the presence of Git references
for each pull request, the existing scripts used to publish the status
of a pull request need to be updated. To limit the confusion, a new route
is created and the existing one is marked as deprecated.

Statuses are attached to a commit, future usages of this could be
to display the status of each commit of a PR or/and to display the
statuses when bronsing the history from the web UI.

This is part of story #11477: have a dedicated references for pullrequests

Change-Id: If93f16d99e48370070535c508d2696ccf63c8876