Fix "updating" spinner for kanban items
Part of story #11543 Use new palette colors as background color in A.D

I took the opportunity to refactor the "time info" clock into a
dedicated component. This brings the following improvements:
- the clock is now fully replaced by an "updating" spinner instead of
css "hacks" to hide it to show a background image on the whole card.
- the time info title is a one-time binding, which means Angular does
not waste time recomputing it for every card every split second during
drag and drop
- Reduced unmaintainable $scope dependencies between kanban-item and
kanban controller (but some are still there)

How to test:
- The clock icon should be horizontally centered with regards to the
accessibility pattern
- The clock icon should be shown for all columns except Backlog
- It should still show a localized message stating the time the item has
been in Kanban and in its current column
- It should be replaced by an "updating" spinner when the card is drag
and dropped or edited
- "Updating" cards keep their background color

Change-Id: Ica83631abd881ababf1957321e46180cedab2a5b