TOTP codes can be validated
This contribution demonstrates that Tuleap is now capable
to validate TOTP codes.

To test, you can access the page /plugins/mfa/enroll
a base32 encoded value is displayed, it is the shared
secret between the client and the server validating the
TOTP codes.

You can generate the TOTP code corresponding to the secret
using the CLI tool oathtool (any other tool will work too)
this way:

    oathtool --totp -b <secret>

You will get a code that you can set in the input field to check
if it is valid or not.

This is part of story #11649: protect my account with a TOTP code

Change-Id: Ib4ecda515000f1de65c716181f4c939cd484e674