request #11672: Do not try to reserve a Git reference when a PR is updated
To allow the realization of story #11477 to be as smooth and as easy as possible
to do and to review, if a Git reference is not reserved when a PR is updated Tuleap
will try to create it.

This case must not happen anymore so the code dealing with this case is basically
dead code.
If somehow this case happens, Tuleap should fail and logs the issue as it indicates
the instance is not an expected state (likely forgeupgrade was not run) and trying
to continue the execution might breaks more things.

Change-Id: I90e30867da119e0e0c9ac9b694b9433718b8740a