request #11675: Updating an artifact through SOAP sets an invalid comment format in database
The verification of the comment format takes into account the type of the
value being verified, a boolean true value is not the same than "text" or
"html" value.
Also, the comment format given to the SOAP API is now taken into account, previously
the value true was always set regardless of what the user chosen.

Change-Id: I35752e5e325a5edec83a8fc7adb330f73d0538bf