Send a webhook at artifact creation / update
This is part of story #11035 trigger webhook on artifact updates

The webhook URL must be stored directly in DB to be sent (the UI to do
this will be done at the very end of the story).

A fake payload content is sent to check that the webhook was correctly
sent. Some web services can do this for you [1]

For the object injection, I did this as best effort. A refactoring will
be possible at the end of the story to better handle this (if
necessary).

The log system is not yet implemented. It will be done in a dedicated
commit to reduce the size and the scope of this commit.

Next commit will introduce the real payload content.

[1] https://requestbin.fullcontact.com/

Change-Id: I37c2d65de7506fd6f4d2717eaa18f5128e0e9826