request #11687: Drop SHA1 polyfill
sha1() is available since PHP 4.3.0 (and should not be used nowadays)
there is no need for this polyfill.
Also, it is expected from implementation of cryptographic functions to
be constant-time, this implementation is not (information is leaked by functions
like chr() and ord() that uses a lookup table and so can be exploited by processor
cache misses).

Change-Id: Ia2dda408f8075c67539aa248aa2e354181e2dab0