Move GitPHP in the Tuleap sources
Minimal changes has been made to the GitPHP sources, everything that
was working before this contribution should continue to work.

A specifity on the blobdiff action that was present when a GET parameter
named jenkins was set to "true" has been removed. The use case for this
seems not to be known anymore and anything relying on the UI will
probably be broken with the upcoming work anyway.

This is part of story #10411: cleanly integrate gitphp into tuleap

Change-Id: I8e83fb69d4e392ec06f09aabca5ecfe1c49be293